fix: Reorganize tests
