+++ tests/to_refactor/python/test_liveness.py
@@ -1,7 +1,7 @@
 import os
 from dotenv import load_dotenv
 
-from tests.python.utils import client
+from tests.to_refactor.fixtures import client
 
 load_dotenv()
 
@@ -16,7 +16,7 @@ def test_calculate_liveness(client):
             + os.getenv("AUTH0_TEST_TOKEN_NO_DOMAIN"),
         },
         data=open(
-            "./tests/python/data/calculate_us_1_data.json",
+            "./tests/to_refactor/python/data/calculate_us_1_data.json",
             "r",
             encoding="utf-8",
         ),
